Fixed tracing for SentryGraphQLHttpMessageHandler when no transaction is set on the scope
