+++ test/Sentry.Tests/SentryGraphQlHttpMessageHandlerTests.cs
@@ -7,7 +7,7 @@ namespace Sentry.Tests;
  * TODO: Find a way to consolidate these tests cleanly.
  */
 
-public class SentryGraphQlHttpMessageHandlerTests
+public class SentryGraphQlHttpMessageHandlerTests : SentryMessageHandlerTests
 {
     private const string ValidQuery = "query getAllNotes { notes { id } }";
     private const string ValidResponse = @"{
@@ -47,13 +47,7 @@ public class SentryGraphQlHttpMessageHandlerTests
     public void ProcessRequest_SetsSpanData()
     {
         // Arrange
-        var hub = Substitute.For<IHub>();
-        var parentSpan = Substitute.For<ISpan>();
-        hub.GetSpan().Returns(parentSpan);
-        var childSpan = Substitute.For<ISpan>();
-        parentSpan.When(p => p.StartChild(Arg.Any<string>()))
-                  .Do(op => childSpan.Operation = op.Arg<string>());
-        parentSpan.StartChild(Arg.Any<string>()).Returns(childSpan);
+        var hub = _fixture.GetHub();
         var sut = new SentryGraphQLHttpMessageHandler(hub, null);
 
         var method = "POST";
@@ -66,9 +60,15 @@ public class SentryGraphQlHttpMessageHandlerTests
 
         // Assert
         returnedSpan.Should().NotBeNull();
-        returnedSpan!.Operation.Should().Be("http.client");
-        returnedSpan.Description.Should().Be($"{method} {url}");
-        returnedSpan.Received(1).SetExtra(OtelSemanticConventions.AttributeHttpRequestMethod, method);
+        using (new AssertionScope())
+        {
+            returnedSpan!.Operation.Should().Be("http.client");
+            returnedSpan.Description.Should().Be($"{method} {url}");
+            returnedSpan.Extra.Should().Contain(kvp =>
+                kvp.Key == OtelSemanticConventions.AttributeHttpRequestMethod &&
+                kvp.Value.ToString() == method
+            );
+        }
     }
 
     // [Theory]
@@ -118,14 +118,13 @@ public class SentryGraphQlHttpMessageHandlerTests
     public void HandleResponse_SetsSpanData()
     {
         // Arrange
-        var hub = Substitute.For<IHub>();
+        var hub = _fixture.GetHub();
         var status = HttpStatusCode.OK;
         var response = new HttpResponseMessage(status);
         var method = "POST";
         var url = "http://example.com/graphql";
         var request = SentryGraphQlTestHelpers.GetRequestQuery(ValidQuery, url);
         response.RequestMessage = request;
-        hub.GetSpan().Returns(new TransactionTracer(hub, "test", "test"));
         var sut = new SentryGraphQLHttpMessageHandler(hub, null);
 
         // Act
