+++ test/Sentry.Tests/SentryHttpMessageHandlerTests.cs
@@ -7,33 +7,8 @@ namespace Sentry.Tests;
  * TODO: Find a way to consolidate these tests cleanly.
  */
 
-public class SentryHttpMessageHandlerTests
+public class SentryHttpMessageHandlerTests: SentryMessageHandlerTests
 {
-    private class Fixture
-    {
-        public readonly SentryOptions Options = new();
-
-        public readonly ISentryClient Client = Substitute.For<ISentryClient>();
-
-        public readonly ISessionManager SessionManager = Substitute.For<ISessionManager>();
-
-        public readonly ISystemClock Clock = Substitute.For<ISystemClock>();
-
-        public Fixture()
-        {
-            Options.Dsn = ValidDsn;
-            Options.TracesSampleRate = 1.0;
-        }
-
-        public Hub GetHub()
-        {
-            var scopeManager = new SentryScopeManager(Options, Client);
-            return new Hub(Options, Client, SessionManager, Clock, scopeManager);
-        }
-    }
-
-    private readonly Fixture _fixture = new();
-
     [Fact]
     public async Task SendAsync_SentryTraceHeaderNotSet_SetsHeader_ByDefault()
     {
